requirements.txt
appdirs==1.4.4
asgiref==3.3.4
bootstrap4==0.1.0
distlib==0.3.1
Django==3.2.3
django-cors-headers==3.7.0
django-sass==1.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
libsass==0.20.1
mysqlclient==2.0.3
psycopg2==2.8.6
PyJWT==2.1.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.2
